HIA-1635 Move redactionConfig
